+++ umh-core/pkg/communicator/pkg/generator/generator.go
@@ -19,13 +19,8 @@ import (
 
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/communicator/pkg/tools/watchdog"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/config"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/config/dataflowcomponentserviceconfig"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/constants"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsm"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsm/agent_monitor"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsm/container"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsm/dataflowcomponent"
-	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/fsm/redpanda"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/logger"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/models"
 	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/sentry"
@@ -34,10 +29,12 @@ import (
 
 const (
 	// Manager name constants
-	containerManagerName = logger.ComponentContainerManager + "_" + constants.DefaultManagerName
-	benthosManagerName   = logger.ComponentBenthosManager + "_" + constants.DefaultManagerName
-	agentManagerName     = logger.ComponentAgentManager + "_" + constants.DefaultManagerName
-	redpandaManagerName  = logger.ComponentRedpandaManager + constants.DefaultManagerName
+	// TODO: clean up these constants
+	containerManagerName         = logger.ComponentContainerManager + "_" + constants.DefaultManagerName
+	benthosManagerName           = logger.ComponentBenthosManager + "_" + constants.DefaultManagerName
+	agentManagerName             = logger.ComponentAgentManager + "_" + constants.DefaultManagerName
+	redpandaManagerName          = logger.ComponentRedpandaManager + constants.DefaultManagerName
+	dataflowcomponentManagerName = constants.DataflowcomponentManagerName
 	// Instance name constants
 	coreInstanceName     = "Core"
 	agentInstanceName    = "agent"
@@ -93,6 +90,32 @@ func (s *StatusCollectorType) getDataFlowComponentData() ([]models.Dfc, error) {
 	return dfcData, nil
 }
 
+// findInstance finds an instance in the system snapshot
+// this is useful if we want to fetch the data from a manager, that always has one instance (e.g., core, agent, container, redpanda)
+// returns nil if the instance is not found
+func findInstance(
+	snap fsm.SystemSnapshot,
+	managerName, instanceName string,
+) (*fsm.FSMInstanceSnapshot, bool) {
+
+	mgr, ok := snap.Managers[managerName]
+	if !ok {
+		return nil, false
+	}
+	inst, ok := mgr.GetInstances()[instanceName]
+	return inst, ok
+}
+
+// findManager finds a manager in the system snapshot
+// returns nil if the manager is not found
+func findManager(
+	snap fsm.SystemSnapshot,
+	managerName string,
+) (fsm.ManagerSnapshot, bool) {
+	mgr, ok := snap.Managers[managerName]
+	return mgr, ok
+}
+
 func (s *StatusCollectorType) GenerateStatusMessage() *models.StatusMessage {
 
 	// Step 1: Get the snapshot
@@ -103,71 +126,35 @@ func (s *StatusCollectorType) GenerateStatusMessage() *models.StatusMessage {
 		return nil
 	}
 
-	// Step 2: Get container data from snapshot via the container manager if available
+	// Step 2: Build the status message
+
+	// --- container (only one instance) ---------------------------------------------------------
 	var containerData models.Container
-
-	if containerManager, exists := snapshot.Managers[containerManagerName]; exists {
-		instances := containerManager.GetInstances()
-
-		if instance, ok := instances[coreInstanceName]; ok {
-			containerData = buildContainerDataFromSnapshot(*instance, s.logger)
-		} else {
-			s.logger.Warn("Core instance not found in container manager",
-				zap.String("instanceName", coreInstanceName))
-			containerData = buildDefaultContainerData()
-		}
-	} else {
-		s.logger.Warn("Container manager not found in system snapshot",
-			zap.String("managerName", containerManagerName))
-
-		containerData = buildDefaultContainerData()
+	contInst, ok := findInstance(snapshot, containerManagerName, coreInstanceName)
+	if ok {
+		containerData = buildContainerDataFromInstanceSnapshot(*contInst, s.logger)
 	}
 
-	// Step 3: Get dataflowcomponent data from snapshot via the dataflowcomponent manager if available
-	dfcData, err := s.getDataFlowComponentData()
-	if err != nil {
-		s.logger.Error("Error getting dataflowcomponent data", zap.Error(err))
-	}
-
-	// Step 4: Get agent data from the agent manager if available
+	// --- agent (only one instance) -------------------------------------------------------------
 	var agentData models.Agent
-	var releaseChannel string
-
-	if agentManager, exists := snapshot.Managers[agentManagerName]; exists {
-		instances := agentManager.GetInstances()
-
-		s.logger.Debug("Agent manager instances",
-			zap.Any("instances", instances))
-
-		if instance, ok := instances[agentInstanceName]; ok {
-			agentData, releaseChannel = buildAgentDataFromSnapshot(*instance, s.logger)
-		} else {
-			s.logger.Warn("Agent instance not found in agent manager",
-				zap.String("instanceName", agentInstanceName),
-				zap.Any("instances", instances))
-			sentry.ReportIssuef(sentry.IssueTypeError, s.logger, "[GenerateStatusMessage] Agent instance not found in agent manager")
-			agentData = models.Agent{}
-			releaseChannel = "n/a"
-		}
+	var agentDataReleaseChannel string
+	agInst, ok := findInstance(snapshot, agentManagerName, agentInstanceName)
+	if ok {
+		agentData, agentDataReleaseChannel = agentFromInstanceSnapshot(agInst, s.logger)
 	}
 
-	// Step 5: Get redpanda data from the redpanda manager if available
+	// --- redpanda (only one instance) -------------------------------------------------------------
 	var redpandaData models.Redpanda
+	rpInst, ok := findInstance(snapshot, redpandaManagerName, redpandaInstanceName)
+	if ok {
+		redpandaData = redpandaFromInstanceSnapshot(rpInst, s.logger)
+	}
 
-	if redpandaManager, exists := snapshot.Managers[redpandaManagerName]; exists {
-		instances := redpandaManager.GetInstances()
-
-		if instance, ok := instances[redpandaInstanceName]; ok {
-			redpandaData, err = buildRedpandaDataFromSnapshot(*instance, s.logger)
-			if err != nil {
-				s.logger.Error("Error building redpanda data", zap.Error(err))
-			}
-		} else {
-			s.logger.Warn("Redpanda instance not found in redpanda manager",
-				zap.String("instanceName", redpandaInstanceName))
-			sentry.ReportIssuef(sentry.IssueTypeError, s.logger, "[GenerateStatusMessage] Redpanda instance not found in redpanda manager")
-			redpandaData = models.Redpanda{}
-		}
+	// --- dfc (multiple instances) ----------------------	---------------------------------------
+	var dfcData []models.Dfc
+	dfcMgr, ok := findManager(snapshot, constants.DataflowcomponentManagerName)
+	if ok {
+		dfcData = dfcsFromManagerSnapshot(dfcMgr, s.logger)
 	}
 
 	// Create the status message
@@ -190,7 +177,7 @@ func (s *StatusCollectorType) GenerateStatusMessage() *models.StatusMessage {
 					Category:      models.Neutral,
 				},
 				Version: "n/a",
-				Channel: releaseChannel,
+				Channel: agentDataReleaseChannel,
 				SupportedFeatures: []string{
 					"custom-dfc",
 					"action-deploy-data-flow-component",
@@ -214,133 +201,6 @@ func (s *StatusCollectorType) GenerateStatusMessage() *models.StatusMessage {
 	return statusMessage
 }
 
-// buildContainerDataFromSnapshot creates container data from a FSM instance snapshot
-func buildContainerDataFromSnapshot(instance fsm.FSMInstanceSnapshot, log *zap.SugaredLogger) models.Container {
-	// Try to get observed state from instance
-	containerData := buildDefaultContainerData()
-
-	// Check if we have actual observedState
-	if instance.LastObservedState != nil {
-		// Try to cast to the right type
-		if snapshot, ok := instance.LastObservedState.(*container.ContainerObservedStateSnapshot); ok {
-			status := snapshot.ServiceInfoSnapshot
-
-			// Create health status
-			containerData.Health = &models.Health{
-				Message:       getHealthMessage(status.OverallHealth),
-				ObservedState: instance.CurrentState,
-				DesiredState:  instance.DesiredState,
-				Category:      status.OverallHealth,
-			}
-
-			// Fill in CPU metrics
-			if status.CPU != nil {
-				containerData.CPU = status.CPU
-				// Ensure health is set
-				if containerData.CPU.Health == nil {
-					containerData.CPU.Health = &models.Health{
-						Message:       getHealthMessage(status.CPUHealth),
-						ObservedState: status.CPUHealth.String(),
-						DesiredState:  models.Active.String(),
-						Category:      status.CPUHealth,
-					}
-				}
-			}
-
-			// Fill in Memory metrics
-			if status.Memory != nil {
-				containerData.Memory = status.Memory
-				// Ensure health is set
-				if containerData.Memory.Health == nil {
-					containerData.Memory.Health = &models.Health{
-						Message:       getHealthMessage(status.MemoryHealth),
-						ObservedState: status.MemoryHealth.String(),
-						DesiredState:  models.Active.String(),
-						Category:      status.MemoryHealth,
-					}
-				}
-			}
-
-			// Fill in Disk metrics
-			if status.Disk != nil {
-				containerData.Disk = status.Disk
-				// Ensure health is set
-				if containerData.Disk.Health == nil {
-					containerData.Disk.Health = &models.Health{
-						Message:       getHealthMessage(status.DiskHealth),
-						ObservedState: status.DiskHealth.String(),
-						DesiredState:  models.Active.String(),
-						Category:      status.DiskHealth,
-					}
-				}
-			}
-
-			// Set hardware info
-			containerData.Hwid = status.Hwid
-			containerData.Architecture = status.Architecture
-		} else {
-			log.Warn("Container observed state is not of expected type")
-		}
-	} else {
-		log.Warn("Container instance has no observed state")
-	}
-
-	return containerData
-}
-
-func buildDataFlowComponentDataFromSnapshot(instance fsm.FSMInstanceSnapshot, log *zap.SugaredLogger) (models.Dfc, error) {
-	dfcData := models.Dfc{}
-
-	// Check if we have actual observedState
-	if instance.LastObservedState != nil {
-		// Try to cast to the right type
-
-		// Create health status based on instance.CurrentState
-		extractHealthStatus := func(state string) models.HealthCategory {
-			switch state {
-			case dataflowcomponent.OperationalStateActive:
-				return models.Active
-			case dataflowcomponent.OperationalStateDegraded:
-				return models.Degraded
-			default:
-				return models.Neutral
-			}
-		}
-
-		// get the metrics from the instance
-		observed, ok := instance.LastObservedState.(*dataflowcomponent.DataflowComponentObservedStateSnapshot)
-		if !ok {
-			err := fmt.Errorf("observed state %T does not match DataflowComponentObservedStateSnapshot", instance.LastObservedState)
-			log.Error(err)
-			return models.Dfc{}, err
-		}
-		serviceInfo := observed.ServiceInfo
-		inputThroughput := float64(0)
-		if serviceInfo.BenthosObservedState.ServiceInfo.BenthosStatus.BenthosMetrics.MetricsState != nil && serviceInfo.BenthosObservedState.ServiceInfo.BenthosStatus.BenthosMetrics.MetricsState.Input.LastCount > 0 {
-			inputThroughput = serviceInfo.BenthosObservedState.ServiceInfo.BenthosStatus.BenthosMetrics.MetricsState.Input.MessagesPerTick / constants.DefaultTickerTime.Seconds()
-		}
-
-		dfcData.Health = &models.Health{
-			Message:       getDataflowHealthMessage(extractHealthStatus(instance.CurrentState), *observed),
-			ObservedState: instance.CurrentState,
-			DesiredState:  instance.DesiredState,
-			Category:      extractHealthStatus(instance.CurrentState),
-		}
-
-		dfcData.Type = "custom" // this is a custom DFC; protocol converters will have a separate fsm
-		dfcData.UUID = dataflowcomponentserviceconfig.GenerateUUIDFromName(instance.ID).String()
-		dfcData.Metrics = &models.DfcMetrics{
-			AvgInputThroughputPerMinuteInMsgSec: inputThroughput,
-		}
-		dfcData.Name = &instance.ID
-	} else {
-		log.Warn("no observed state found for dataflowcomponent", zap.String("instanceID", instance.ID))
-		return models.Dfc{}, fmt.Errorf("no observed state found for dataflowcomponent")
-	}
-
-	return dfcData, nil
-}
-
 // getHealthMessage returns an appropriate message based on health category
 func getHealthMessage(health models.HealthCategory) string {
 	switch health {
@@ -353,171 +213,4 @@ func getHealthMessage(health models.HealthCategory) string {
 	default:
 		return "Component status unknown"
 	}
-}
-
-// getDataflowHealthMessage returns an appropriate message for dataflow components
-func getDataflowHealthMessage(health models.HealthCategory, serviceInfo dataflowcomponent.DataflowComponentObservedStateSnapshot) string {
-	switch health {
-	case models.Active:
-		return "Component is operating normally"
-	case models.Degraded:
-		if serviceInfo.ServiceInfo.BenthosObservedState.ServiceInfo.BenthosStatus.BenthosDegradedLog.Content != "" {
-			return fmt.Sprintf("Component is degraded because of a bad log entry. If the problem persists, please check the logs for more information. Log entry: [ %s ] %s",
-				serviceInfo.ServiceInfo.BenthosObservedState.ServiceInfo.BenthosStatus.BenthosDegradedLog.Timestamp.String(),
-				serviceInfo.ServiceInfo.BenthosObservedState.ServiceInfo.BenthosStatus.BenthosDegradedLog.Content)
-		}
-		return "Component stopped working"
-	case models.Neutral:
-		return "Component status is neutral"
-	default:
-		return "Component status unknown"
-	}
-}
-
-// buildDefaultContainerData creates default container data when no real data is available
-func buildDefaultContainerData() models.Container {
-	return models.Container{
-		Health: &models.Health{
-			Message:       "Container status unknown",
-			ObservedState: "unknown",
-			DesiredState:  "running",
-			Category:      models.Neutral,
-		},
-		CPU: &models.CPU{
-			Health: &models.Health{
-				Message:       "CPU status unknown",
-				ObservedState: "unknown",
-				DesiredState:  "normal",
-				Category:      models.Neutral,
-			},
-			TotalUsageMCpu: 0,
-			CoreCount:      0,
-		},
-		Disk: &models.Disk{
-			Health: &models.Health{
-				Message:       "Disk status unknown",
-				ObservedState: "unknown",
-				DesiredState:  "normal",
-				Category:      models.Neutral,
-			},
-			DataPartitionUsedBytes:  0,
-			DataPartitionTotalBytes: 0,
-		},
-		Memory: &models.Memory{
-			Health: &models.Health{
-				Message:       "Memory status unknown",
-				ObservedState: "unknown",
-				DesiredState:  "normal",
-				Category:      models.Neutral,
-			},
-			CGroupUsedBytes:  0,
-			CGroupTotalBytes: 0,
-		},
-		Hwid:         "unknown",
-		Architecture: models.ArchitectureAmd64,
-	}
-}
-
-// buildRedpandaDataFromSnapshot creates redpanda data from a FSM instance snapshot
-// the instance will give us currentState, desiredState, etc. and the last observed state
-func buildRedpandaDataFromSnapshot(instance fsm.FSMInstanceSnapshot, log *zap.SugaredLogger) (models.Redpanda, error) {
-	redpandaData := models.Redpanda{}
-
-	// Check if we have actual observedState
-	if instance.LastObservedState != nil {
-
-		// Step 1: Assemble Health Information
-
-		// extract the current and desired state from the instance
-		currentState := instance.CurrentState
-		desiredState := instance.DesiredState
-
-		// Now derive the health category from the current state
-		healthCategory := models.Neutral
-		switch currentState {
-		case redpanda.OperationalStateActive:
-			healthCategory = models.Active
-		case redpanda.OperationalStateDegraded:
-			healthCategory = models.Degraded
-		}
-
-		// Now derive the health message from the health category
-		healthMessage := getHealthMessage(healthCategory)
-
-		// Now assemble the health status
-		redpandaData.Health = &models.Health{
-			Message:       healthMessage,
-			ObservedState: currentState,
-			DesiredState:  desiredState,
-			Category:      healthCategory,
-		}
-
-		// Step 2: Assemble Redpanda Metrics (AvgIncomingThroughputPerMinuteInMsgSec, AvgOutgoingThroughputPerMinuteInMsgSec)
-		// Try to cast to the right type
-		observedState, ok := instance.LastObservedState.(*redpanda.RedpandaObservedStateSnapshot)
-		if !ok {
-			err := fmt.Errorf("observed state %T does not match RedpandaObservedStateSnapshot", instance.LastObservedState)
-			log.Error(err)
-			return models.Redpanda{}, err
-		}
-
-		// Now fetch the last observed service info
-		observedStateServiceInfo := observedState.ServiceInfoSnapshot
-
-		// Now calculate the throughput metrics if available
-		if observedStateServiceInfo.RedpandaStatus.RedpandaMetrics.MetricsState != nil {
-			redpandaData.AvgIncomingThroughputPerMinuteInMsgSec = float64(observedStateServiceInfo.RedpandaStatus.RedpandaMetrics.MetricsState.Input.BytesPerTick) / constants.DefaultTickerTime.Seconds()
-			redpandaData.AvgOutgoingThroughputPerMinuteInMsgSec = float64(observedStateServiceInfo.RedpandaStatus.RedpandaMetrics.MetricsState.Output.BytesPerTick) / constants.DefaultTickerTime.Seconds()
-		}
-
-	} else {
-		log.Warn("no observed state found for redpanda", zap.String("instanceID", instance.ID))
-		return models.Redpanda{}, fmt.Errorf("no observed state found for redpanda")
-	}
-
-	return redpandaData, nil
-}
-
-// buildAgentDataFromSnapshot creates agent data from a FSM instance snapshot
-func buildAgentDataFromSnapshot(instance fsm.FSMInstanceSnapshot, log *zap.SugaredLogger) (models.Agent, string) {
-	agentData := models.Agent{}
-	releaseChannel := "n/a"
-	// Check if we have actual observedState
-	if instance.LastObservedState != nil {
-		// Try to cast to the right type
-		if snapshot, ok := instance.LastObservedState.(*agent_monitor.AgentObservedStateSnapshot); ok {
-			// Ensure all fields are valid before accessing
-			if snapshot.ServiceInfoSnapshot.Location == nil {
-				log.Warn("Agent location data is nil")
-				agentData = models.Agent{
-					Location: map[int]string{0: "Unknown location"},
-				}
-			} else {
-				agentData = models.Agent{
-					Location: snapshot.ServiceInfoSnapshot.Location,
-				}
-			}
-			// build the health status
-			agentData.Health = &models.Health{
-				Message:       getHealthMessage(snapshot.ServiceInfoSnapshot.OverallHealth),
-				ObservedState: instance.CurrentState,
-				DesiredState:  instance.DesiredState,
-				Category:      snapshot.ServiceInfoSnapshot.OverallHealth,
-			}
-			// Check if Release is nil before accessing its properties
-			if snapshot.ServiceInfoSnapshot.Release != nil {
-				releaseChannel = snapshot.ServiceInfoSnapshot.Release.Channel
-			} else {
-				log.Warn("Agent release data is nil, defaulting to stable")
-				releaseChannel = "n/a"
-			}
-		} else {
-			log.Warn("Agent observed state is not of expected type")
-			sentry.ReportIssuef(sentry.IssueTypeError, log, "[buildAgentDataFromSnapshot] Agent observed state is not of expected type")
-		}
-	} else {
-		log.Warn("Agent instance has no observed state")
-	}
-
-	return agentData, releaseChannel
 }
